+++ src/main/java/arm11/ReadFunctions.java
@@ -8,12 +8,10 @@ public class ReadFunctions {
     list.add(InstructionFactory.createLabel(readLabel))
         .add(InstructionFactory.createPush(ARM11Registers.LR))
         .add(InstructionFactory.createMove(ARM11Registers.R1,
-                ARM11Registers.R0))
-        .add(InstructionFactory.createLoad(ARM11Registers.R0,
-                                           formatterLabel))
-        .add(InstructionFactory.createAdd(ARM11Registers.R0,
-                                          ARM11Registers.R0,
-                                          new Immediate((long) 4)))
+            ARM11Registers.R0))
+        .add(InstructionFactory.createLoad(ARM11Registers.R0, formatterLabel))
+        .add(InstructionFactory.createAdd(ARM11Registers.R0, ARM11Registers.R0,
+            new Immediate(4L)))
         .add(InstructionFactory.createBranchLink(new Label("scanf")));
   }
